Refactor SearchSuggestionFilter into separate class
