+++ lib/l10n/app_de.arb
@@ -265,17 +265,18 @@
   "pageSearchSuggestionsTitle": "Vorschläge",
   "pageSearchSuggestionsEmpty": "Nichts zu sehen",
   "pageSearchSuggestionsTooltipFilter": "Filter",
-  "pageSearchSuggestionsFilterMinimumRating": "Mindeste Bewertung",
-  "pageSearchSuggestionsFilterFeatures": "Features",
-  "pageSearchSuggestionsFilterFeaturesSelectTooltip": "Auswählen",
-  "pageSearchSuggestionsFilterFeaturesDeselectTooltip": "Abwählen ",
-  "pageSearchSuggestionsFilterDeviation": "Maximale Abweichung von ausgewählter Zeit",
-  "pageSearchSuggestionsFilterDeviationHours": "Stunden",
-  "pageSearchSuggestionsFilterSorting": "Sortierung",
-  "pageSearchSuggestionsFilterResetToDefault": "Zurücksetzen",
-  "pageSearchSuggestionsSortingRelevance": "Relevanz",
-  "pageSearchSuggestionsSortingTime": "Zeit",
-  "pageSearchSuggestionsSortingPrice": "Preis",
+
+  "searchSuggestionsFilterMinimumRating": "Mindeste Bewertung",
+  "searchSuggestionsFilterFeatures": "Features",
+  "searchSuggestionsFilterFeaturesSelectTooltip": "Auswählen",
+  "searchSuggestionsFilterFeaturesDeselectTooltip": "Abwählen ",
+  "searchSuggestionsFilterDeviation": "Maximale Abweichung von ausgewählter Zeit",
+  "searchSuggestionsFilterDeviationHours": "Stunden",
+  "searchSuggestionsFilterSorting": "Sortierung",
+  "searchSuggestionsFilterResetToDefault": "Zurücksetzen",
+  "searchSuggestionsSortingRelevance": "Relevanz",
+  "searchSuggestionsSortingTime": "Zeit",
+  "searchSuggestionsSortingPrice": "Preis",
 
   "pageRidesTitle": "Mitfahrten",
   "pageRidesTooltipSearchRide": "Fahrt suchen",
