+++ lib/l10n/app_en.arb
@@ -390,17 +390,18 @@
   "pageSearchSuggestionsTitle": "Search Suggestions",
   "pageSearchSuggestionsEmpty": "Nothing to see",
   "pageSearchSuggestionsTooltipFilter": "Filter",
-  "pageSearchSuggestionsFilterMinimumRating": "Minimum rating",
-  "pageSearchSuggestionsFilterFeatures": "Features",
-  "pageSearchSuggestionsFilterFeaturesSelectTooltip": "Select",
-  "pageSearchSuggestionsFilterFeaturesDeselectTooltip": "Deselect",
-  "pageSearchSuggestionsFilterDeviation": "Maximum deviation from selected time",
-  "pageSearchSuggestionsFilterDeviationHours": "hours",
-  "pageSearchSuggestionsFilterSorting": "Sorting",
-  "pageSearchSuggestionsFilterResetToDefault": "Reset to default",
-  "pageSearchSuggestionsSortingRelevance": "Relevance",
-  "pageSearchSuggestionsSortingTime": "Time",
-  "pageSearchSuggestionsSortingPrice": "Price",
+
+  "searchSuggestionsFilterFeaturesSelectTooltip": "Select",
+  "searchSuggestionsFilterFeatures": "Features",
+  "searchSuggestionsFilterFeaturesDeselectTooltip": "Deselect",
+  "searchSuggestionsFilterDeviation": "Maximum deviation from selected time",
+  "searchSuggestionsFilterMinimumRating": "Minimum rating",
+  "searchSuggestionsFilterDeviationHours": "hours",
+  "searchSuggestionsFilterSorting": "Sorting",
+  "searchSuggestionsFilterResetToDefault": "Reset to default",
+  "searchSuggestionsSortingRelevance": "Relevance",
+  "searchSuggestionsSortingTime": "Time",
+  "searchSuggestionsSortingPrice": "Price",
 
   "pageRidesTitle": "Rides",
   "pageRidesTooltipSearchRide": "Search ride",
